CarIdentifier removing toInt

+++ app/src/main/java/com/tomerpacific/caridentifier/composable/CarInformation.kt
@@ -39,12 +39,12 @@ import com.tomerpacific.caridentifier.concatenateCarMakeAndModel
 import com.tomerpacific.caridentifier.model.CarDetails
 import com.tomerpacific.caridentifier.model.MainViewModel
 
-private val COLOR_YELLOW = Color(0xFFFDB000.toInt())
-private val COLOR_BLUE = Color(0xFF1250FF.toInt())
-private val COLOR_GREEN = Color(0xFF32A852.toInt())
-private val COLOR_RED = Color(0xFFFF0000.toInt())
-private val COLOR_BROWN = Color(0xFFD19669.toInt())
-private val COLOR_BLACK = Color(0xFF000000.toInt())
+private val COLOR_YELLOW = Color(0xFFFDB000)
+private val COLOR_BLUE = Color(0xFF1250FF)
+private val COLOR_GREEN = Color(0xFF32A852)
+private val COLOR_RED = Color(0xFFFF0000)
+private val COLOR_BROWN = Color(0xFFD19669)
+private val COLOR_BLACK = Color(0xFF000000)
 
 @Composable
 fun Details(mainViewModel: MainViewModel) {
